Change Double values to BigDecimal (stored as String)

+++ app/src/main/java/pi/restaurant/management/ui/fragments/discounts/AbstractModifyDiscountFragment.kt
@@ -63,7 +63,7 @@ abstract class AbstractModifyDiscountFragment : AbstractModifyItemFragment() {
             assignedDiscounts = discountsViewModel.createDiscounts(code, assignedNumber, availableNumber),
             usedDiscounts = discountsViewModel.createDiscounts(code, usedNumber, availableNumber + assignedNumber),
             type = binding.spinnerType.selectedItemId.toInt(),
-            amount = binding.editTextAmount.text.toString().toDouble(),
+            amount = binding.editTextAmount.text.toString(),
             expirationDate = Date() //TODO Zrobić
         )
